requirements.txt
# Core RAG pipeline
PyMuPDF>=1.24.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
requests>=2.31.0

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Client SDK
httpx>=0.25.0
